Allow for setting the Next server IP in the DHCP header
In order to PXE boot a baremetal server using the OVN DHCP server we
need to allow users to set the "next-server" (siaddr) [0] field in the
DHCP header.

While investigating this issue by comparing the DHCPOFFER and DHCPACK
packets sent my dnsmasq and OVN we saw that the "next-server" field
was the problem for OVN, without it PXE booting was timing out while
fetching the iPXE image from the TFTP server (see the bugzilla ticket
below for reference).

To confirm this problem we created a bogus patch hardcoding the TFTP
address in the siaddr of the DHCP header (see the discussion in the
maillist below) and with this in place we were able to deploy a
baremetal node using the OVN DHCP end-to-end.

This patch is a proper implementation that creates a new DHCP
configuration option called "next_server" to allow users to set this
field dynamically. This patch uses the DHCP code 253 which is a unsed
code for DHCP specification as this is not a normal DHCP option but a
special use case in OVN.

[0]
https://github.com/openvswitch/ovs/blob/9dd3031d2e0e9597449e95428320ccaaff7d8b3d/lib/dhcp.h#L42

Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2083629
Reported-at:
https://mail.openvswitch.org/pipermail/ovs-discuss/2022-May/051821.html
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>
Signed-off-by: Numan Siddique <numans@ovn.org>