test: Make vtep test pass more consistently.
The "vtep: 3 HVs, 1 VIFs/HV, 1 GW, 1 LS" test fails consistently on PPC.
The test expects a flow with the following to be present:

is_chassis_resident("cr-lrp1") || is_chassis_resident("cr-lrp2")

However, on PPC, the test fails because the flow has the following
instead:

is_chassis_resident("cr-lrp2") || is_chassis_resident("cr-lrp1")

In other words, the fields are reversed. This is because the ordering of
the ports in ovn-northd can differ depending on architecture. This is a
case where the northd code is correct, but the test is overly
restrictive.

To fix this, the test has been broken into two parts. First, we check
the overall flow structure to ensure that there are two
"is_chassis_resident" fields present. Then we can break that flow down
further and sort the fields so that they will be in a predictable order.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Dumitru Ceara <dceara@redhat.com>