northd.c: Add option to enable MAC learning on localnet
The localnet is excluded from MAC learning for scale
reason. However there might be a valid workflow
when yo uwant to enable the learning and benefit
for that for HW offload. Add option called
'localnet_learn_fdb' to LSP, which will enable/disable
the learning. Setting it as disabled by default.

Reported-at: https://bugzilla.redhat.com/2070529
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>