treewide: Disable storing of other_config in external_ids
The update of having other_config instead of external_ids
for configuration has been in place for some time.
Version that released with this change was 20.06, that
is plenty of time so we don't have to keep the compatibility
anymore. Remove the updates of external_ids it will be stored
only in other_config from now on.

Signed-off-by: Ales Musil <amusil@redhat.com>
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>