controller: Fix wrong controller action definition
The controller action had a wrong userdata_len due to
missing subtraction of previous data len. The pointer to oc
was not moved correctly, as result it could point to wrong
memory. In order to fix that expose encode_controller_op and
use it instead of manually creating the action. At the same
time use stack buffer for the ofpbuf which should be large
enough to hold related data without any heap allocations.

Fixes: e52c2451 ("physical.c: Fix bug of wrong use in vm migration")
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Han Zhou <hzhou@ovn.org>