northd: don't add drop lflow if LB VIP matches LRP IP
If it is needed to create Load Balancer within LR with VIP, which matches
any of LR's LRP IP, there is no need to create SNAT entry.  Now such
traffic destined to LRP IP is not dropped.

With this patch a drop lflow with match=(ipX.dst == {IP}) is not added to
lr_in_ip_input stage if LRP IP matches associated with this LR LB VIP.

Tests are added as well.

Signed-off-by: Vladislav Odintsov <odivlad@gmail.com>
Signed-off-by: Numan Siddique <numans@ovn.org>