northd: Accumulate more database updates before processing.
northd doesn't process changes incrementally, so it makes sense to
accumulate more database updates and process them in bulk, so we can
cover everything in a single recompute.

ovsdb-server has a mechanism to start accumulating changes if the
client doesn't receive them fast enough, but it relies on the receive
buffer size, which is a few hundreds of KB on a typical system.
Unfortunately, that is enough to queue up several hundreds of small
updates, and it takes northd a lot of time to process them if poll
intervals are large, receiving at most 50 messages on each iteration
(half of which are updates for a _Server database).

Calling ovsdb_idl_run() as long as something changes.  This allows to
quickly process large bursts of database updates.  For example, it
takes only 30-40 seconds for 'ovn-nbctl --wait=hv sync' to finish on
a 500-node cluster after the startup phase of the density-heavy
ovn-heater test, instead of 6-8 minutes without this change.

500 ms seems like a reasonable hard limit to avoid spinning for too
long if the database is changed constantly at a fast pace.

Very long polling is also logged at INFO level to notify users.
Not using WARN or higher because it may happen under normal conditions,
e.g. on the initial connection to a large database or another type
of a single large update.  Other notable polling attempts are logged
at debug level.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Han Zhou <hzhou@ovn.org>