northd: Retrieve load balancer options only once.
Options like 'add_route' and 'neighbor_responder' are looked up
in the smap for each load balancer for each datapath and it takes
significant amount of time.  Checking them only once to speed up
processing.  'skip_snat' is not that critical, but it's better to
get all of them in a single place.

Also using enum for the neighbor responder mode to avoid costly
string comparison on a hot path.

Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Han Zhou <hzhou@ovn.org>