controller: Fix QoS for ports with undetected or too low link speed.
If netdev_set_qos() is called without specifying max-rate, it will
use reported link speed of that interface instead as a ceil.

If OVS libraries are not able to detect the link speed, the 100 Mbps
will be used.  Also, TAP interfaces do report 10 Mbps as their link
speed by default.

In both cases all the traffic will be unnecessary limited to just
100 or 10 Mpbs regardless of the 'qos_max_rate' configured for that
port in OVN databases, because maximum rate of a single queue can not
be higher than a total maximum rate of a QoS type.

Fix that by always specifying the max-rate as a maximum supported
value - (2^32 - 1) * 8 bits per second.  This way netdev_set_qos()
will not need to check the link speed or guess a value.

Fixes: 8dda482b8059 ("Check and allocate free qdisc queue id for ports with qos parameters")
Reported-at: https://bugzilla.redhat.com/2136716
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>