ci: Update jobs to use numbers instead of test flags
To prevent some jobs not running after CI scripts updates
use numbers instead of the test flags. This still allows us
to use parallelization, but without worrying about skipping
some tests by mistake.

For "test" suites use 3 parts, with 1500 tests in mind.
That should give as additional space for future tests.
Currently, there is ~1200 tests. For sanitizers use 5 parts
instead, as they are slower in general.

For "system-test" use 3 parts, with 300 tests in mind.
Currently, there is ~200 tests.

In the end this patch reduces the number of jobs by 8 to 20,
which is ok as there is a limit of 20 running in parallel [0].

[0] https://docs.github.com/en/actions/learn-github-actions/usage-limits-billing-and-administration#usage-limits
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
(cherry picked from commit 03449fa72baf7e1a824bf2660d3d1675f03217de)