ovn-controller: Fix releasing wrong vif
When binding an OVS interface to a logical port, if external_ids:iface-id is
set within the same transaction as adding the interface to a bridge,
ovn-controller is usually initially notified of ovs interface change with
ofport == 0.
If the lport was already bound to a different interface, the ofport=0
notification caused an unexpected release of the interface.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Acked-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: Numan Siddique <numans@ovn.org>
(cherry picked from commit a2f80ce2436306bf4938ef7c7925ece3d6cd3e20)