CI: Update Cirrus CI to run ARM jobs
The Cirrus CI script was configured to run FreeBSD job,
however it seems that this job didn't run for several years.
Update the .cirrus.yml to run ARM jobs instead. As first
phase only the regular "test" for GCC and Clang.

We cannot run sanitizers for now as they are very slow
on Cirrus CI and just timeout without ever finishing.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>